refactor: separate hooks file for Web3Provider
