+++ src/components/Web3Provider/index.tsx
@@ -1,44 +1,11 @@
 import { Web3ReactProvider } from '@web3-react/core'
-import { Connector } from '@web3-react/types'
-import { BACKFILLABLE_WALLETS, getConnectorForWallet, gnosisSafe, injected, network, useConnectors } from 'connectors'
-import { ReactNode, useEffect } from 'react'
-import { useAppSelector } from 'state/hooks'
+import { ReactNode } from 'react'
 
-import { isMobile } from '../../utils/userAgent'
-
-const connect = async (connector: Connector) => {
-  try {
-    if (connector.connectEagerly) {
-      await connector.connectEagerly()
-    } else {
-      await connector.activate()
-    }
-  } catch (error) {
-    console.debug(`web3-react eager connection error: ${error}`)
-  }
-}
+import { useConnectors, useEagerlyConnect } from './hooks'
 
 export default function Web3Provider({ children }: { children: ReactNode }) {
-  const selectedWalletBackfilled = useAppSelector((state) => state.user.selectedWalletBackfilled)
-  const selectedWallet = useAppSelector((state) => state.user.selectedWallet)
-
-  const connectors = useConnectors(selectedWallet)
-
-  const isMetaMask = !!window.ethereum?.isMetaMask
-
-  useEffect(() => {
-    connect(gnosisSafe)
-    connect(network)
-
-    if (isMobile && isMetaMask) {
-      injected.activate()
-    } else if (selectedWallet) {
-      connect(getConnectorForWallet(selectedWallet))
-    } else if (!selectedWalletBackfilled) {
-      BACKFILLABLE_WALLETS.map(getConnectorForWallet).forEach(connect)
-    }
-    // The dependency list is empty so this is only run once on mount
-  }, []) // eslint-disable-line react-hooks/exhaustive-deps
+  useEagerlyConnect()
+  const connectors = useConnectors()
 
   return <Web3ReactProvider connectors={connectors}>{children}</Web3ReactProvider>
 }
